move different modules clean code
