+++ src/main/java/org/jargus/alert/client/AlertSystemClient.java
@@ -1,7 +1,6 @@
-package org.jargus.collect.client;
+package org.jargus.alert.client;
 
 import org.jargus.collect.model.MetricInfo;
-import org.jargus.common.model.Metric;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
