+++ src/main/java/org/jargus/alert/configuration/AlertSystemClientConfiguration.java
@@ -1,4 +1,4 @@
-package org.jargus.collect.configuration;
+package org.jargus.alert.configuration;
 
 /**
  * @author Bazhov N.S.
