+++ src/main/java/org/jargus/analyze/storage/MetricsAnalysisRuleValidator.java
@@ -1,7 +1,6 @@
-package org.jargus.collect.storage;
+package org.jargus.analyze.storage;
 
 import org.jargus.collect.model.MetricInfo;
-import org.jargus.common.model.Metric;
 
 /**
  * @author Bazhov N.S.
