+++ src/main/java/org/jargus/collect/mapper/ModuleResponseMapper.java
@@ -2,7 +2,6 @@ package org.jargus.collect.mapper;
 
 import org.jargus.collect.model.MetricInfo;
 import org.jargus.common.dto.ModuleResponse;
-import org.jargus.common.model.Metric;
 
 /**
  * @author Bazhov N.S.
