+++ src/main/java/org/jargus/collect/service/InternalDatabaseServiceImpl.java
@@ -3,18 +3,17 @@ package org.jargus.collect.service;
 import lombok.RequiredArgsConstructor;
 import org.jargus.collect.model.DatabaseMetricsRequestParams;
 import org.jargus.collect.model.MetricInfo;
-import org.jargus.common.dao.InternalDatabaseDao;
+import org.jargus.database.dao.TsStorageClient;
 
 /**
  * @author Bazhov N.S.
  */
 @RequiredArgsConstructor
 public class InternalDatabaseServiceImpl implements InternalDatabaseService {
-    private final InternalDatabaseDao internalDatabaseDao;
+    private final TsStorageClient tsStorageClient;
 
     @Override
     public MetricInfo getMetrics(DatabaseMetricsRequestParams databaseMetricsRequestParams) {
-        internalDatabaseDao.getMetrics(null);
         return null;
     }
 }
